WIP pull data from the graph
Pulling data from the graph is faster since it is indexed mostly the way
we need it. The one issue is that the sync events are not easily
accessible from the graph. Trying a few different things, it looks like
getting the synced reserves would be easiest by tracking them by pulling
mints and burns and checking them using the hourly pair query.

TODOs:
[] Pull mints
[] Pull burns
[] Iterate through mints, burns, and swaps to keep reserves updated
[] check reserves against hourly data to ensure correctness